+++ src/analyzer/include/Hodo1Hit.hh
@@ -40,7 +40,7 @@ public:
   bool   Calculate( bool tdc_flag = true );
   bool   IsCalculated( void ) const { return m_is_calculated; }
   int    GetNumOfHit(int sel=0) const { return sel==0 ? m_multi_hit_l : m_multi_hit_t; };
-  double GetA( int n=0 )   const { if (n<m_a.size()) return m_a.at(n); 
+  double GetA( int n=0 )   const { if (n<static_cast<int>(m_a.size())) return m_a.at(n);
                                    else return -9999.;}
   double GetT( int n=0 )   const { return m_t.at(n); }
   double GetCT( int n=0 )  const { return m_ct.at(n); }
@@ -74,7 +74,7 @@ public:
 
   virtual
   int DetectorId( void ) const { return m_raw->DetectorId(); }
-  virtual 
+  virtual
   int PlaneId( void )    const { return m_raw->PlaneId(); }
   virtual
   int SegmentId( void )  const { return m_raw->SegmentId(); }
